requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
asttokens==3.0.0
asyncio==3.4.3
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
decorator==5.2.1
diskcache==5.6.3
distro==1.9.0
Django==5.2.1
exceptiongroup==1.3.0
executing==2.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ipython==8.36.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
llama_cpp_python==0.3.9
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
numpy==2.2.6
openai==1.82.0
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.51
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
python-dotenv==1.1.0
regex==2024.11.6
requests==2.32.3
sniffio==1.3.1
sqlparse==0.5.3
stack-data==0.6.3
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
wcwidth==0.2.13